requirements.txt
Flask
Flask-Admin
SQLAlchemy==1.3.15
Flask-SQLAlchemy
Flask-Security-too
Flask-Migrate
sqlalchemy_utils